+++ lib/grape-swagger/entity/parser.rb
@@ -3,21 +3,31 @@ module GrapeSwagger
     class Parser
       attr_reader :model
       attr_reader :endpoint
+      attr_reader :attribute_parser
 
       def initialize(model, endpoint)
         @model = model
         @endpoint = endpoint
+        @attribute_parser = AttributeParser.new(endpoint)
       end
 
       def call
-        parameters = model.root_exposures.each_with_object({}) do |value, memo|
-          memo[value.attribute] = value.send(:options)
-        end
-
-        parse_grape_entity_params(parameters)
+        parse_grape_entity_params(extract_params(model))
       end
 
       private
+
+      def extract_params(exposure)
+        exposure.root_exposures.each_with_object({}) do |value, memo|
+          if value.for_merge && (value.respond_to?(:entity_class) || value.respond_to?(:using_class_name))
+            entity_class = value.respond_to?(:entity_class) ? value.entity_class : value.using_class_name
+
+            memo.merge!(extract_params(entity_class))
+          else
+            memo[value.attribute] = value.send(:options)
+          end
+        end
+      end
 
       def parse_grape_entity_params(params, parent_model = nil)
         return unless params
@@ -25,49 +35,20 @@ module GrapeSwagger
         params.each_with_object({}) do |(entity_name, entity_options), memo|
           next if entity_options.fetch(:documentation, {}).fetch(:in, nil).to_s == 'header'
 
-          entity_name = entity_options[:as] if entity_options[:as]
+          final_entity_name = entity_options.fetch(:as, entity_name)
           documentation = entity_options[:documentation]
-          entity_model = model_from(entity_options)
 
-          if entity_model
-            name = endpoint.nil? ? entity_model.to_s.demodulize : endpoint.send(:expose_params_from_model, entity_model)
-            memo[entity_name] = entity_model_type(name, entity_options)
-          elsif entity_options[:nesting]
-            memo[entity_name] = parse_nested(entity_name, entity_options, parent_model)
-          else
-            memo[entity_name] = data_type_from(entity_options)
-            next unless documentation
+          memo[final_entity_name] = if entity_options[:nesting]
+                                      parse_nested(entity_name, entity_options, parent_model)
+                                    else
+                                      attribute_parser.call(entity_options)
+                                    end
 
-            memo[entity_name][:default] = documentation[:default] if documentation[:default]
-
-            if (values = documentation[:values])
-              memo[entity_name][:enum] = values if values.is_a?(Array)
-            end
-
-            if documentation[:is_array]
-              memo[entity_name] = {
-                type: :array,
-                items: memo.delete(entity_name)
-              }
-            end
-          end
-
-          if documentation
-            memo[entity_name][:read_only] = documentation[:read_only].to_s == 'true' if documentation[:read_only]
-            memo[entity_name][:description] = documentation[:desc] if documentation[:desc]
-            memo[entity_name].merge!(documentation[:documentation]) if documentation[:documentation]
-          end
+        next unless documentation
+          memo[final_entity_name][:readOnly] = documentation[:read_only].to_s == 'true' if documentation[:read_only]
+          memo[final_entity_name][:description] = documentation[:desc] if documentation[:desc]
+          memo[final_entity_name].merge!(documentation[:documentation]) if documentation[:documentation]
         end
-      end
-
-      def model_from(entity_options)
-        model = entity_options[:using] if entity_options[:using].present?
-
-        if could_it_be_a_model?(entity_options[:documentation])
-          model ||= entity_options[:documentation][:type]
-        end
-
-        model
       end
 
       def parse_nested(entity_name, entity_options, parent_model = nil)
@@ -81,6 +62,8 @@ module GrapeSwagger
           memo[value.attribute] = value.send(:options)
         end
 
+        required = required_params(params)
+
         properties = parse_grape_entity_params(params, nested_entity)
         is_a_collection = entity_options[:documentation].is_a?(Hash) &&
                           entity_options[:documentation][:type].to_s.casecmp('array').zero?
@@ -88,71 +71,29 @@ module GrapeSwagger
         if is_a_collection
           {
             type: :array,
-            items: {
+            items: with_required({
               type: :object,
               properties: properties
-            },
-            description: entity_options[:desc] || ''
+            }, required)
           }
         else
-          {
+          with_required({
             type: :object,
-            properties: properties,
-            description: entity_options[:desc] || ''
-          }
+            properties: properties
+          }, required)
         end
       end
 
-      def could_it_be_a_model?(value)
-        return false if value.nil?
-        direct_model_type?(value[:type]) || ambiguous_model_type?(value[:type])
+      def required_params(params)
+        params.select { |_, options| options.fetch(:documentation, {}).fetch(:required, false) }
+              .map { |(key, options)| [options.fetch(:as, key), options] }
+              .map(&:first)
       end
 
-      def direct_model_type?(type)
-        type.to_s.include?('Entity') || type.to_s.include?('Entities')
-      end
-
-      def ambiguous_model_type?(type)
-        type &&
-          type.is_a?(Class) &&
-          !GrapeSwagger::DocMethods::DataType.primitive?(type.name.downcase) &&
-          !type == Array
-      end
-
-      def data_type_from(documentation)
-        documented_type = documentation[:type]
-        documented_type ||= (documentation[:documentation] && documentation[:documentation][:type])
-
-        data_type = GrapeSwagger::DocMethods::DataType.call(documented_type)
-
-        document_data_type(documentation[:documentation], data_type)
-      end
-
-      def document_data_type(documentation, data_type)
-        type = if GrapeSwagger::DocMethods::DataType.primitive?(data_type)
-                 data = GrapeSwagger::DocMethods::DataType.mapping(data_type)
-                 { type: data.first, format: data.last }
-               else
-                 { type: data_type }
-               end
-        type[:format] = documentation[:format] if documentation && documentation.key?(:format)
-
-        type
-      end
-
-      def entity_model_type(name, entity_options)
-        if entity_options[:documentation] && entity_options[:documentation][:is_array]
-          {
-            'type' => 'array',
-            'items' => {
-              '$ref' => "#/definitions/#{name}"
-            }
-          }
-        else
-          {
-            '$ref' => "#/definitions/#{name}"
-          }
-        end
+      def with_required(hash, required)
+        return hash if required.empty?
+        hash[:required] = required
+        hash
       end
     end
   end